Change MessageType from u16 to u32
Breaking wire protocol change: MessageType field increased from 2 bytes
to 4 bytes in LP packets. This future-proofs the message type space and
aligns with other u32 fields.

Changes:
- message.rs: #[repr(u32)], from_u32(), to_u32()
- error.rs: InvalidMessageType(u32)
- codec.rs: All serialization/deserialization updated to 4-byte msg_type
  - Cleartext parsing: inner_bytes[4..8], content at [8..]
  - AEAD parsing: decrypted[4..8], content at [8..]
  - Serialization: 4 bytes for message type

+++ common/nym-lp/src/error.rs
@@ -26,7 +26,7 @@ pub enum LpError {
     InvalidPacketFormat(String),
 
     #[error("Invalid message type: {0}")]
-    InvalidMessageType(u16),
+    InvalidMessageType(u32),
 
     #[error("Payload too large: {0}")]
     PayloadTooLarge(usize),
